requirements.txt
python-dotenv==1.0.0
openai==0.28.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.4.2 